requirements.txt
asgiref==3.5.1
certifi==2021.10.8
charset-normalizer==2.0.12
cloudinary==1.29.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==4.0.4
django-cloudinary-storage==0.3.0
django-cors-headers==3.11.0
django-environ==0.8.1
django-heroku==0.3.1
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
Pillow==9.1.0
psycopg2==2.9.3
PyJWT==2.3.0
pytz==2022.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
static3==0.7.0
tzdata==2022.1
urllib3==1.26.9
whitenoise==6.0.0
